Refactoring common imports into Trait to clean-up test files.
