+++ src/test/scala/io/opentargets/etl/backend/EtlSparkUnitTest.scala
@@ -1,0 +1,11 @@
+package io.opentargets.etl.backend
+
+import org.scalatest.PrivateMethodTester
+import org.scalatest.flatspec.AnyFlatSpecLike
+import org.scalatest.matchers.must.Matchers
+
+trait EtlSparkUnitTest
+  extends AnyFlatSpecLike
+    with PrivateMethodTester
+    with SparkSessionSetup
+    with Matchers
